refactor(cleanup): split config errors and execute-run tests
