+++ src/frame_compare/error_context.py
@@ -1,0 +1,69 @@
+"""Core error context types for Frame Compare."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+type JSONValue = str | int | float | bool | None | list["JSONValue"] | dict[str, "JSONValue"]
+type ErrorDetails = dict[str, JSONValue]
+
+
+@dataclass(frozen=True, slots=True)
+class ErrorContext:
+    """Structured error information."""
+
+    code: str
+    name: str
+    message: str
+    details: ErrorDetails | None = None
+    hint: str | None = None
+    cause: BaseException | None = None
+
+    def to_dict(self) -> dict[str, JSONValue]:
+        result: dict[str, JSONValue] = {
+            "code": self.code,
+            "name": self.name,
+            "message": self.message,
+        }
+        if self.hint:
+            result["hint"] = self.hint
+        if self.details:
+            result["details"] = self.details
+        return result
+
+
+class FrameCompareError(Exception):
+    """Base exception for all Frame Compare errors."""
+
+    def __init__(self, context: ErrorContext) -> None:
+        self.context = context
+        super().__init__(context.message)
+
+    @property
+    def code(self) -> str:
+        return self.context.code
+
+    @property
+    def name(self) -> str:
+        return self.context.name
+
+    @property
+    def hint(self) -> str | None:
+        return self.context.hint
+
+    def __str__(self) -> str:
+        base = f"[{self.code}] {self.context.message}"
+        if self.hint:
+            base += f"\nHint: {self.hint}"
+        return base
+
+    def __repr__(self) -> str:
+        return f"{self.__class__.__name__}({self.context!r})"
+
+
+__all__ = [
+    "ErrorContext",
+    "ErrorDetails",
+    "FrameCompareError",
+    "JSONValue",
+]
