+++ src/frame_compare/services/tmdb_lookup.py
@@ -13,6 +13,93 @@ TMDB_IMAGE_ORIGINAL_BASE_URL = "https://image.tmdb.org/t/p/original"
 TMDB_KEY_REGEX = re.compile(r"^[0-9a-fA-F]{32}$")
 
 
+def _build_search_params(parsed: ParsedMetadata, api_key: str) -> dict[str, str | int]:
+    params: dict[str, str | int] = {
+        "api_key": api_key,
+        "query": parsed.title,
+        "language": "en-US",
+        "page": 1,
+        "include_adult": "false",
+    }
+    if parsed.year:
+        params["year"] = parsed.year
+    return params
+
+
+async def _request_tmdb_search(
+    params: dict[str, str | int],
+    config: MetadataConfig,
+    client: httpx.AsyncClient,
+) -> httpx.Response:
+    return await client.get(
+        TMDB_API_URL,
+        params=params,
+        timeout=config.timeout_seconds,
+    )
+
+
+def _raise_for_tmdb_response(response: httpx.Response) -> None:
+    if response.status_code == 401:
+        raise TmdbError("Invalid API key")
+    if response.status_code == 429:
+        raise TmdbRateLimitedError()
+    if response.status_code >= 500:
+        raise TmdbError(f"TMDB service error: {response.status_code}")
+
+    response.raise_for_status()
+
+
+def _decode_tmdb_json(response: httpx.Response) -> dict[str, Any]:
+    return cast(dict[str, Any], response.json())
+
+
+def _tmdb_image_url(path: str | None) -> str | None:
+    if not path:
+        return None
+    return f"{TMDB_IMAGE_ORIGINAL_BASE_URL}{path}"
+
+
+def _tmdb_result_year(item: dict[str, Any], media_type: str) -> int:
+    if media_type == "movie":
+        year_str = str(item.get("release_date", ""))[:4]
+    else:
+        year_str = str(item.get("first_air_date", ""))[:4]
+
+    return int(year_str) if year_str.isdigit() else 0
+
+
+def _map_tmdb_result(item: dict[str, Any]) -> TmdbMetadata | None:
+    media_type_raw = item.get("media_type", "movie")
+    if media_type_raw == "movie":
+        media_type = "movie"
+    elif media_type_raw == "tv":
+        media_type = "tv"
+    else:
+        return None
+
+    return TmdbMetadata(
+        tmdb_id=int(item["id"]),
+        title=str(item.get("title") or item.get("name", "Unknown")),
+        original_title=str(item.get("original_title") or item.get("original_name", "Unknown")),
+        year=_tmdb_result_year(item, media_type),
+        media_type=media_type,
+        poster_url=_tmdb_image_url(cast(str | None, item.get("poster_path"))),
+        backdrop_url=_tmdb_image_url(cast(str | None, item.get("backdrop_path"))),
+    )
+
+
+def _map_tmdb_results(data: dict[str, Any]) -> list[TmdbMetadata]:
+    results_raw = cast(list[dict[str, Any]], data.get("results", []))
+    mapped_results: list[TmdbMetadata] = []
+
+    for item in results_raw:
+        result = _map_tmdb_result(item)
+        if result is not None:
+            mapped_results.append(result)
+
+    return mapped_results
+
+
 def is_valid_tmdb_api_key(api_key: str) -> bool:
     """Return whether a TMDB API key matches the API v3 key format."""
     return TMDB_KEY_REGEX.fullmatch(api_key) is not None
@@ -30,33 +117,12 @@ async def _search_tmdb(
     if not is_valid_tmdb_api_key(config.api_key):
         raise TmdbError("Invalid API key format")
 
-    params: dict[str, str | int] = {
-        "api_key": config.api_key,
-        "query": parsed.title,
-        "language": "en-US",
-        "page": 1,
-        "include_adult": "false",
-    }
-    if parsed.year:
-        params["year"] = parsed.year
+    params = _build_search_params(parsed, config.api_key)
 
     try:
-        response = await client.get(
-            TMDB_API_URL,
-            params=params,
-            timeout=config.timeout_seconds,
-        )
-
-        if response.status_code == 401:
-            raise TmdbError("Invalid API key")
-        if response.status_code == 429:
-            raise TmdbRateLimitedError()
-        if response.status_code >= 500:
-            raise TmdbError(f"TMDB service error: {response.status_code}")
-
-        response.raise_for_status()
-        data = cast(dict[str, Any], response.json())
-
+        response = await _request_tmdb_search(params, config, client)
+        _raise_for_tmdb_response(response)
+        data = _decode_tmdb_json(response)
     except httpx.TimeoutException as e:
         raise TmdbError("Request timed out") from e
     except httpx.RequestError as e:
@@ -70,47 +136,7 @@ async def _search_tmdb(
         # Avoid leaking request details (e.g., query params) via exception stringification.
         raise TmdbError("Unexpected error during TMDB lookup") from e
 
-    results_raw = cast(list[dict[str, Any]], data.get("results", []))
-    mapped_results: list[TmdbMetadata] = []
-
-    for item in results_raw:
-        media_type_raw = item.get("media_type", "movie")
-        if media_type_raw == "movie":
-            media_type = "movie"
-        elif media_type_raw == "tv":
-            media_type = "tv"
-        else:
-            continue
-
-        # Year extraction
-        year_str = ""
-        if media_type == "movie":
-            year_str = str(item.get("release_date", ""))[:4]
-        else:
-            year_str = str(item.get("first_air_date", ""))[:4]
-
-        year = int(year_str) if year_str.isdigit() else 0
-
-        def _get_url(path: str | None) -> str | None:
-            if not path:
-                return None
-            return f"{TMDB_IMAGE_ORIGINAL_BASE_URL}{path}"
-
-        mapped_results.append(
-            TmdbMetadata(
-                tmdb_id=int(item["id"]),
-                title=str(item.get("title") or item.get("name", "Unknown")),
-                original_title=str(
-                    item.get("original_title") or item.get("original_name", "Unknown")
-                ),
-                year=year,
-                media_type=media_type,
-                poster_url=_get_url(cast(str | None, item.get("poster_path"))),
-                backdrop_url=_get_url(cast(str | None, item.get("backdrop_path"))),
-            )
-        )
-
-    return mapped_results
+    return _map_tmdb_results(data)
 
 
 search_tmdb = _search_tmdb
