+++ tests/orchestration/test_execution_phase_plan.py
@@ -1,0 +1,98 @@
+"""Tests for execute_run phase planning and request override contracts."""
+
+from __future__ import annotations
+
+from datetime import datetime
+from pathlib import Path
+
+from frame_compare.config.schema import ConfigSchema, OverlayMode, TonemapPreset
+from frame_compare.orchestration.coordinator import RunDependencies, RunRequest
+from frame_compare.orchestration.execution import build_execution_phase_plan
+from frame_compare.orchestration.types import (
+    ExecutionState,
+    MetadataPrefetch,
+    PrepState,
+    RunArtifacts,
+)
+from frame_compare.utils.types import WorkspacePaths
+
+from .execute_run_helpers import FakeFFmpegRunner, clip_state
+
+
+def test_build_execution_phase_plan_preserves_align_boundary_and_progress_total(
+    tmp_path: Path,
+) -> None:
+    workspace = WorkspacePaths(
+        root=tmp_path,
+        input_dir=tmp_path / "comparison_videos",
+        run_dir=None,
+        screenshots_dir=tmp_path / "screenshots",
+        generated_dir=tmp_path / "generated",
+        config_dir=tmp_path / "config",
+        config_file=tmp_path / "config" / "config.toml",
+    )
+
+    prep = PrepState(
+        workspace=workspace,
+        config=ConfigSchema(),
+        input_videos=[
+            tmp_path / "ref.mkv",
+            tmp_path / "comp_a.mkv",
+            tmp_path / "comp_b.mkv",
+        ],
+        clips=[
+            clip_state(tmp_path / "ref.mkv", label="Reference"),
+            clip_state(tmp_path / "comp_a.mkv", label="Encode 1"),
+            clip_state(tmp_path / "comp_b.mkv", label="Encode 2"),
+        ],
+        artifacts=RunArtifacts(),
+        metadata_prefetch=MetadataPrefetch(None, False),
+        preflight_warnings=[],
+        preflight_duration=0.0,
+        load_sources_start=datetime.now(),
+    )
+
+    plan = build_execution_phase_plan(
+        request=RunRequest(root=tmp_path),
+        deps=RunDependencies(ffmpeg_runner=FakeFFmpegRunner()),
+        prep=prep,
+        state=ExecutionState(artifacts=prep.artifacts),
+    )
+
+    assert [phase.name for phase in plan.before_align] == ["frame_plan", "analyze", "align"]
+    assert [phase.name for phase in plan.after_align] == [
+        "render",
+        "metadata",
+        "dovi",
+        "publish",
+        "report",
+    ]
+
+    align_phase = next(phase for phase in plan.before_align if phase.name == "align")
+    assert align_phase.progress_total == 1
+
+
+def test_run_request_cli_config_overrides_capture_runtime_override_contract(tmp_path: Path) -> None:
+    request = RunRequest(
+        root=tmp_path,
+        input_dir=tmp_path / "comparison_videos",
+        tm_preset=TonemapPreset.FILMIC,
+        tm_target_nits=203,
+        overlay_mode=OverlayMode.DIAGNOSTIC,
+        frame_count=12,
+        seed=123,
+        no_upload=True,
+        force_interactive_alignment=True,
+    )
+
+    overrides = request.cli_config_overrides()
+
+    assert overrides.input_dir == tmp_path / "comparison_videos"
+    assert overrides.tm_preset == TonemapPreset.FILMIC
+    assert overrides.tm_target_nits == 203
+    assert overrides.tm_curve is None
+    assert overrides.frame_count == 12
+    assert overrides.seed == 123
+    assert overrides.overlay_mode == OverlayMode.DIAGNOSTIC
+    assert overrides.no_upload is True
+    assert overrides.force_interactive_alignment is True
